Refactor Category & VotesDate testing logic
