+++ front/src/__tests__/pages/Category.test.tsx
@@ -1,7 +1,9 @@
 import { expect, vi } from 'vitest'
-import { render, screen, fireEvent, waitFor } from '../test-utils'
+import { render, renderHook, screen, fireEvent, waitFor } from '../test-utils'
 import { Category } from '../../pages'
 import { TAuthContext, useAuth } from '../../context/AuthProvider'
+import { useSortByDate } from '../../hooks/useSortByDate'
+import { useSortByVotes } from '../../hooks/useSortByVotes'
 
 vi.mock('react-router-dom', async () => {
   const actual: Record<number, unknown> = await vi.importActual(
@@ -141,4 +143,68 @@ it('status filter widget appears for users who are logged in', () => {
 
   const statusFilterWidget = screen.getByTestId('status-filter')
   expect(statusFilterWidget).toBeInTheDocument()
+})
+
+it('sorts resources by date in descending order', () => {
+  const items = [
+    { id: 1, date: '2023-11-01' },
+    { id: 2, date: '2023-10-30' },
+    { id: 3, date: '2023-10-28' }
+  ]
+  render(<Category />)
+
+  fireEvent.click(screen.getByText('Data'))
+
+  const { result } = renderHook(() => useSortByDate(items, 'date', 'desc'))
+
+  expect(result.current.sortedItems).toEqual([
+    { id: 1, date: '2023-11-01' },
+    { id: 2, date: '2023-10-30' },
+    { id: 3, date: '2023-10-28' },
+  ])
+})
+
+it('sorts resources by votes in ascending order', () => {
+  const votes = [
+    {
+      id: 'resource1',
+      title: 'Resource 1',
+      voteCount: {
+        downvote: 1,
+        upvote: 2,
+        total: 3,
+        userVote: 1,
+      },
+    },
+    {
+      id: 'resource2',
+      title: 'Resource 2',
+      voteCount: {
+        downvote: 0,
+        upvote: 0,
+        total: 0,
+        userVote: 0,
+      },
+    },
+    {
+      id: 'resource3',
+      title: 'Resource 3',
+      voteCount: {
+        downvote: 0,
+        upvote: 0,
+        total: 7,
+        userVote: 0,
+      },
+    },
+  ]
+  render(<Category />)
+
+  fireEvent.click(screen.getByText('Vots'))
+
+  const { result } = renderHook(() => useSortByVotes(votes, 'asc'))
+  const sortedResources = result.current.sortedVotes
+
+  const voteCounts = sortedResources.map((vote) => vote.voteCount.total)
+
+  expect(voteCounts).toEqual([0, 3, 7])
 })
